# Conflicts:
#	src/main/java/com/IceHockeyLeague/LeagueManager/Coach/CoachStats.java
#	src/main/java/com/IceHockeyLeague/LeagueManager/Coach/ICoachStats.java
#	src/main/java/com/IceHockeyLeague/LeagueManager/LeagueManagerFactory.java
#	src/main/java/com/IceHockeyLeague/LeagueManager/Player/IPlayerStats.java
#	src/main/java/com/IceHockeyLeague/LeagueManager/Player/PlayerStats.java
#	src/main/java/com/IceHockeyLeague/LeagueManager/Team/ITeam.java
#	src/main/java/com/IceHockeyLeague/LeagueManager/Team/Team.java
#	src/main/java/com/IceHockeyLeague/LeagueManager/Team/TeamStrengthCalculator.java
#	src/test/java/com/IOTest/CommandLineInputTest.java
#	src/test/java/com/IceHockeyLeagueTest/LeagueFileHandlerTest/LeagueFileValidatorTest.java
#	src/test/java/com/IceHockeyLeagueTest/LeagueFileHandlerTest/LeagueJsonMock.java
#	src/test/java/com/IceHockeyLeagueTest/LeagueManagerTest/CoachTest/CoachDBMock.java
#	src/test/java/com/IceHockeyLeagueTest/LeagueManagerTest/LeagueCreatorTest.java
#	src/test/java/com/IceHockeyLeagueTest/LeagueManagerTest/ManagerTest/ManagerDBMock.java
#	src/test/java/com/IceHockeyLeagueTest/LeagueManagerTest/PlayerTest/PlayerDBMock.java
#	src/test/java/com/IceHockeyLeagueTest/LeagueManagerTest/PlayerTest/PlayerStatsTest.java
#	src/test/java/com/IceHockeyLeagueTest/LeagueManagerTest/PlayerTest/PlayerTest.java
#	src/test/java/com/IceHockeyLeagueTest/LeagueManagerTest/PlayerTest/TeamPlayerDBMock.java
#	src/test/java/com/IceHockeyLeagueTest/LeagueManagerTest/TeamTest/TeamDBMock.java
#	src/test/java/com/IceHockeyLeagueTest/LeagueManagerTest/TestLeagueManagerFactory.java

+++ src/main/java/com/IceHockeyLeague/LeagueManager/Coach/ICoachStats.java
@@ -1,16 +1,16 @@
 package com.IceHockeyLeague.LeagueManager.Coach;
 
 public interface ICoachStats {
-    float getSkating();
-    void setSkating(float value);
+    double getSkating();
+    void setSkating(double value);
 
-    float getShooting();
-    void setShooting(float value);
+    double getShooting();
+    void setShooting(double value);
 
-    float getChecking();
-    void setChecking(float value);
+    double getChecking();
+    void setChecking(double value);
 
-    float getSaving();
-    void setSaving(float value);
+    double getSaving();
+    void setSaving(double value);
 
 }
